kaito-push-3hAM

+++ lib/data/models/contribution_statistics/contribution_statistics_model.dart
@@ -31,7 +31,7 @@ class ContributionStatisticsModel {
 class MembersModel {
   String? name;
   int? contribution;
-  int? percentage;
+  double? percentage;
 
   MembersModel({this.name, this.contribution, this.percentage});
