+++ lib/presentation/screens/order_confirmation/order_confirmation_screen.dart
@@ -47,7 +47,7 @@ class _OrderConfirmationScreenState extends State<OrderConfirmationScreen> {
   bool _isOrdering = false;
   House? house;
   late StreamSubscription<OrderState> _stateSubscription;
-   int priceOfHouseType = 0;
+  int priceOfHouseType = 0;
   Building? selectedBuilding;
 
   @override
@@ -61,11 +61,11 @@ class _OrderConfirmationScreenState extends State<OrderConfirmationScreen> {
 
   Future<void> _fetchPriceByHouseType() async {
     context.read<ServicePriceBloc>().add(
-      GetServicePrice(
-        houseId: house?.id ?? '',
-        serviceId: widget.orderDetails.service.id ?? '',
-      ),
-    );
+          GetServicePrice(
+            houseId: house?.id ?? '',
+            serviceId: widget.orderDetails.service.id ?? '',
+          ),
+        );
   }
 
   void _placeOrder(BuildContext context) {
@@ -105,7 +105,6 @@ class _OrderConfirmationScreenState extends State<OrderConfirmationScreen> {
     widget.orderDetails.houseTypeId = house!.houseTypeId;
   }
 
-
   int _calculateTotalPrice() {
     int total = 0;
 
@@ -128,7 +127,7 @@ class _OrderConfirmationScreenState extends State<OrderConfirmationScreen> {
       total += priceOfHouseType;
     }
 
-    if(widget.orderDetails.emergencyRequest) {
+    if (widget.orderDetails.emergencyRequest) {
       total += 30000; // Add emergency request fee
     }
 
@@ -165,17 +164,18 @@ class _OrderConfirmationScreenState extends State<OrderConfirmationScreen> {
             listener: (context, state) {
               if (state is OrderCreated) {
                 context.read<TransactionBloc>().add(
-                  SaveTransactionEvent(
-                    CreateTransaction(
-                      walletId: selectedWalletId,
-                      paymentMethodId: '15890b1a-f5a6-42c3-8f37-541029189722',
-                      amount: '0',
-                      note: 'Thanh toán dịch vụ',
-                      orderId: state.order.id,
-                      serviceType: 0,
-                    ),
-                  ),
-                );
+                      SaveTransactionEvent(
+                        CreateTransaction(
+                          walletId: selectedWalletId,
+                          paymentMethodId:
+                              '15890b1a-f5a6-42c3-8f37-541029189722',
+                          amount: '0',
+                          note: 'Thanh toán dịch vụ',
+                          orderId: state.order.id,
+                          serviceType: 0,
+                        ),
+                      ),
+                    );
                 showDialog(
                   context: context,
                   barrierDismissible: false,
@@ -216,7 +216,7 @@ class _OrderConfirmationScreenState extends State<OrderConfirmationScreen> {
                         DetailRowWidget(
                           title: 'Thời gian',
                           value:
-                          '${widget.orderDetails.timeSlot.startTime} - ${widget.orderDetails.timeSlot.endTime}',
+                              '${widget.orderDetails.timeSlot.startTime} - ${widget.orderDetails.timeSlot.endTime}',
                           icon: Icons.access_time,
                         ),
                       ],
@@ -230,10 +230,10 @@ class _OrderConfirmationScreenState extends State<OrderConfirmationScreen> {
                         children: widget.orderDetails.option
                             .map(
                               (option) => _buildOptionItem(
-                            title: option.name ?? '',
-                            price: option.price ?? 0,
-                          ),
-                        )
+                                title: option.name ?? '',
+                                price: option.price ?? 0,
+                              ),
+                            )
                             .toList(),
                       ),
                     ),
@@ -245,10 +245,10 @@ class _OrderConfirmationScreenState extends State<OrderConfirmationScreen> {
                         children: widget.orderDetails.extraService
                             .map(
                               (service) => _buildOptionItem(
-                            title: service.name ?? '',
-                            price: service.price ?? 0,
-                          ),
-                        )
+                                title: service.name ?? '',
+                                price: service.price ?? 0,
+                              ),
+                            )
                             .toList(),
                       ),
                     ),
@@ -303,7 +303,7 @@ class _OrderConfirmationScreenState extends State<OrderConfirmationScreen> {
 
                   return _buildPaymentOption(
                     title: walletTitle ?? '',
-                    balance: wallet.balance ?? 0,
+                    balance: (wallet.balance ?? 0).toInt(), // ✅ fix here
                     icon: Icons.account_balance,
                     isSelected: isSelected,
                     onTap: () {
